checked database apis
