+++ lib/connectDb.js
@@ -1,6 +1,9 @@
 import mongoose from "mongoose"; 
 
-const connectDB = async () => {
+const connectToDB = async () => {
+    if (mongoose.connection.readyState === 1) {
+    return;
+  }
     try {
       const conn = await mongoose.connect(process.env.MONGODB_URL || '', {
         useNewUrlParser: true,
@@ -12,4 +15,4 @@ const connectDB = async () => {
     }
   }
 
-export default connectDB; 
+export default connectToDB;
