+++ models/user/user.js
@@ -26,4 +26,4 @@ userSchema.set('toJSON', {
   }
 });
 
-export const User = mongoose.model('User', userSchema);
+export const User = mongoose.models.User || mongoose.model('User', userSchema);
